+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/events.cljs
@@ -1,6 +1,6 @@
 (ns sixsq.nuvla.ui.deployment-dialog.events
   (:require
-    [re-frame.core :refer [dispatch reg-event-db reg-event-fx]]
+    [re-frame.core :refer [dispatch inject-cofx reg-event-db reg-event-fx]]
     [sixsq.nuvla.ui.cimi-api.effects :as cimi-api-fx]
     [sixsq.nuvla.ui.data.spec :as data-spec]
     [sixsq.nuvla.ui.deployment-dialog.spec :as spec]
@@ -18,24 +18,26 @@
      ::cimi-api-fx/delete [(:id deployment) #()]}))
 
 
-(reg-event-db
+(reg-event-fx
   ::set-credentials
-  (fn [db [_ credentials]]
-    (when (= 1 (count credentials))
-      (dispatch [::set-selected-credential (first credentials)]))
-
-    (assoc db ::spec/credentials credentials
-              ::spec/credentials-loading? false)))
+  (fn [{{:keys [::spec/selected-credential-id] :as db} :db} [_ {credentials :resources}]]
+    (let [only-one-cred (= (count credentials) 1)
+          first-cred    (when only-one-cred (first credentials))
+          select-cred?  (and only-one-cred
+                             (not= selected-credential-id (:id first-cred)))]
+      (cond-> {:db (assoc db ::spec/credentials credentials
+                             ::spec/credentials-loading? false)}
+              select-cred? (assoc :dispatch [::set-selected-credential (first credentials)])))))
 
 
 (reg-event-fx
   ::get-data-records-for-cred
   (fn [{{:keys [::spec/deployment
                 ::data-spec/time-period-filter
-                ::spec/infra-service-filter
+                ::spec/cloud-filter
                 ::data-spec/content-type-filter] :as db} :db} _]
     (let [filter        (general-utils/join-and time-period-filter
-                                                infra-service-filter
+                                                cloud-filter
                                                 content-type-filter)
           selected-keys (map keyword (::data-spec/selected-data-set-ids db))
           datasets-map  (select-keys (::data-spec/data-records-by-data-set db) selected-keys)
@@ -57,18 +59,84 @@
 
 
 (reg-event-fx
+  ::set-job-check-cred
+  (fn [{{:keys [::spec/job-check-cred-id
+                ::spec/selected-infra-service]
+         :as   db} :db} [_ {:keys [id return-code] :as resource}]]
+    (when (= job-check-cred-id id)
+      (cond-> {:db (cond-> (assoc db ::spec/check-cred resource)
+                           return-code (assoc ::spec/job-check-cred-id nil
+                                              ::spec/check-cred-loading? false))}
+              return-code (assoc :dispatch [::refresh-credentials])
+              (nil? return-code) (assoc :dispatch-later
+                                        [{:ms 3000 :dispatch [::get-job-check-cred]}])))))
+
+
+(reg-event-fx
+  ::get-job-check-cred
+  (fn [{{:keys [::spec/job-check-cred-id]} :db} _]
+    {::cimi-api-fx/get [job-check-cred-id #(dispatch [::set-job-check-cred %])]}))
+
+
+(reg-event-fx
+  ::set-check-cred-job-id
+  (fn [{:keys [db]} [_ job-id]]
+    {:db             (assoc db ::spec/job-check-cred-id job-id)
+     :dispatch-later [{:ms 3000 :dispatch [::get-job-check-cred]}]}))
+
+
+(reg-event-fx
   ::set-selected-credential
   (fn [{{:keys [::spec/deployment
-                ::spec/data-step-active?] :as db} :db} [_ {credential-id :id :as credential}]]
-    (cond-> {:db (assoc db ::spec/selected-credential credential
-                           ::spec/deployment (assoc deployment :parent credential-id))}
-            data-step-active? (assoc :dispatch [::get-data-records-for-cred]))))
+                ::spec/data-step-active?] :as db} :db} [_ {:keys [id] :as credential}]]
+    (let [check-cred? (utils/credential-need-check? credential)]
+      (cond-> {:db (assoc db ::spec/selected-credential-id id
+                             ::spec/deployment (assoc deployment :parent id)
+                             ::spec/check-cred-loading? check-cred?
+                             ::spec/check-cred nil)}
+
+              data-step-active? (assoc :dispatch [::get-data-records-for-cred])
+              check-cred? (assoc ::cimi-api-fx/operation
+                                 [id
+                                  "check"
+                                  #(dispatch [::set-check-cred-job-id (:location %)])])))))
 
 
 (reg-event-db
   ::set-active-step
   (fn [db [_ active-step]]
     (assoc db ::spec/active-step active-step)))
+
+
+(reg-event-fx
+  ::set-selected-infra-service
+  (fn [{:keys [db]} [_ infra-service]]
+    {:db       (assoc db ::spec/selected-infra-service infra-service
+                         ::spec/check-cred nil
+                         ::spec/check-cred-loading? false)
+     :dispatch [::get-credentials]}))
+
+
+(reg-event-fx
+  ::set-infra-services
+  (fn [{:keys [db]} [_ {infra-services :resources}]]
+    (cond-> {:db (assoc db ::spec/infra-services infra-services
+                           ::spec/infra-services-loading? false)}
+            (= (count infra-services) 1) (assoc :dispatch [::set-selected-infra-service
+                                                           (first infra-services)]))))
+
+
+(reg-event-fx
+  ::get-infra-services
+  (fn [{:keys [db]} [_ filter]]
+    {:db                  (assoc db ::spec/infra-services-loading? true
+                                    ::spec/infra-services nil
+                                    ::spec/selected-infra-service nil)
+     ::cimi-api-fx/search [:infrastructure-service
+                           {:filter filter,
+                            :select "id, name, description, subtype"
+                            :order  "name:asc,id:asc"}
+                           #(dispatch [::set-infra-services %])]}))
 
 
 (reg-event-db
@@ -82,48 +150,58 @@
   ::get-deployment
   (fn [{:keys [db]} [_ id]]
     {:db               (assoc db ::spec/deployment {:id id})
-     ::cimi-api-fx/get [id #(dispatch [::set-deployment %])]}))
+     ::cimi-api-fx/get [id #(let [module-subtype (get-in % [:module :subtype])
+                                  is-kubernetes? (= module-subtype "application_kubernetes")
+                                  filter         (if is-kubernetes?
+                                                   "subtype='kubernetes'"
+                                                   "subtype='swarm'")]
+                              (dispatch [::get-infra-services filter])
+                              (dispatch [::set-deployment %]))]}))
 
 
 (reg-event-fx
   ::create-deployment
   (fn [{{:keys [::spec/deployment] :as db} :db} [_ id first-step do-not-open-modal?]]
     (when (= :data first-step)
-      (dispatch [::get-data-records-by-cred]))
+      (dispatch [::get-data-records]))
     (let [data              {:module {:href id}}
           old-deployment-id (:id deployment)
-          on-success        #(do
-                               (dispatch [::get-credentials])
-                               (dispatch [::get-deployment (:resource-id %)]))]
+          on-success        #(dispatch [::get-deployment (:resource-id %)])]
       (cond->
         {:db               (assoc db ::spec/loading-deployment? true
                                      ::spec/deployment nil
-                                     ::spec/selected-credential nil
+                                     ::spec/selected-credential-id nil
                                      ::spec/deploy-modal-visible? (not (boolean do-not-open-modal?))
                                      ::spec/active-step (or first-step :data)
                                      ::spec/data-step-active? (= first-step :data)
-                                     ::spec/infra-service-filter nil
-                                     ::spec/selected-infra-service nil
-                                     ::spec/infra-services nil
-                                     ::spec/data-infra-services nil)
+                                     ::spec/cloud-filter nil
+                                     ::spec/selected-cloud nil
+                                     ::spec/cloud-infra-services nil
+                                     ::spec/data-clouds nil)
          ::cimi-api-fx/add [:deployment data on-success]}
         old-deployment-id (assoc ::cimi-api-fx/delete [old-deployment-id #() :on-error #()])))))
 
 
 (reg-event-fx
   ::get-credentials
-  (fn [{{:keys [::spec/selected-infra-service] :as db} :db} _]
-    (let [search-creds-callback #(dispatch [::set-credentials (get % :resources [])])]
-      {:db (assoc db ::spec/credentials-loading? true
-                     ::spec/credentials nil
-                     ::spec/selected-credential nil)
-       ::cimi-api-fx/search
-           [:credential
-            {:select "id, name, description, created, subtype"
-             :filter (general-utils/join-and
-                       (when selected-infra-service
-                         (str "parent='" selected-infra-service "'"))
-                       (str "subtype='infrastructure-service-swarm'"))} search-creds-callback]})))
+  (fn [{db :db} _]
+    {:db       (assoc db ::spec/credentials-loading? true
+                         ::spec/credentials nil
+                         ::spec/selected-credential-id nil)
+     :dispatch [::refresh-credentials]}))
+
+
+(reg-event-fx
+  ::refresh-credentials
+  (fn [{{:keys [::spec/selected-infra-service]} :db} _]
+    (when-let [infra-id (:id selected-infra-service)]
+      {::cimi-api-fx/search
+       [:credential
+        {:select "id, name, description, created, subtype, last-check, status"
+         :filter (general-utils/join-and
+                   (str "parent='" infra-id "'")
+                   (str "subtype='infrastructure-service-swarm'"))}
+        #(dispatch [::set-credentials %])]})))
 
 
 (reg-event-fx
@@ -165,24 +243,23 @@
 
 
 (reg-event-db
-  ::set-infra-services
+  ::set-cloud-infra-services
   (fn [db [_ {infra-services :resources}]]
-    (assoc db ::spec/infra-services (into {} (map (juxt :id identity) infra-services)))))
+    (assoc db ::spec/cloud-infra-services (into {} (map (juxt :id identity) infra-services)))))
 
 
-(defn set-infra-service-and-filter
+(defn set-data-infra-service-and-filter
   [db infra-service]
-  (dispatch [::get-credentials])
-  (assoc db ::spec/selected-infra-service infra-service
-            ::spec/infra-service-filter (str "infrastructure-service='" infra-service "'")))
+  (assoc db ::spec/selected-cloud infra-service
+            ::spec/cloud-filter (str "infrastructure-service='" infra-service "'")))
 
 
 (reg-event-fx
-  ::get-infra-services
+  ::get-cloud-infra-services
   (fn [_ [_ filter]]
     {::cimi-api-fx/search [:infrastructure-service
                            {:filter filter, :select "id, name, description, subtype"}
-                           #(dispatch [::set-infra-services %])]}))
+                           #(dispatch [::set-cloud-infra-services %])]}))
 
 
 (reg-event-fx
@@ -193,19 +270,17 @@
           infra-services (map :key buckets)
           filter         (apply general-utils/join-or (map #(str "id='" % "'") infra-services))]
 
-      {:db       (cond-> (assoc db ::spec/data-infra-services buckets)
-                         (= 1 (count infra-services)) (set-infra-service-and-filter
+      {:db       (cond-> (assoc db ::spec/data-clouds buckets)
+                         (= 1 (count infra-services)) (set-data-infra-service-and-filter
                                                         (first infra-services)))
-       :dispatch [::get-infra-services filter]})))
+       :dispatch [::get-cloud-infra-services filter]})))
 
 
 (reg-event-fx
-  ::get-data-records-by-cred
+  ::get-data-records
   (fn [{{:keys [::data-spec/time-period-filter
-                ::data-spec/content-type-filter
-                ::data-spec/credentials] :as db} :db} _]
-    (let [filter (general-utils/join-and time-period-filter
-                                         content-type-filter)]
+                ::data-spec/content-type-filter] :as db} :db} _]
+    (let [filter (general-utils/join-and time-period-filter content-type-filter)]
       {:db db
        ::cimi-api-fx/search
            [:data-record {:filter      filter
@@ -217,4 +292,4 @@
 (reg-event-db
   ::set-infra-service-filter
   (fn [db [_ cloud]]
-    (set-infra-service-and-filter db cloud)))
+    (set-data-infra-service-and-filter db cloud)))