+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/utils.cljs
@@ -1,4 +1,6 @@
-(ns sixsq.nuvla.ui.deployment-dialog.utils)
+(ns sixsq.nuvla.ui.deployment-dialog.utils
+  (:require [sixsq.nuvla.ui.utils.general :as general-utils]
+            [sixsq.nuvla.ui.utils.time :as time]))
 
 
 (defn kw->str
@@ -32,3 +34,36 @@
   (reduce entry-reducer {} dataset-map))
 
 
+(defn credential-last-check-ago
+  [{:keys [last-check] :as credential} locale]
+  (some-> last-check time/parse-iso8601 (time/ago locale)))
+
+
+(defn credential-status-valid
+  [{:keys [status] :as credential}]
+  (some-> status (= "VALID")))
+
+
+(defn credential-is-outdated?
+  [{:keys [last-check] :as credential}]
+  (boolean
+    (or (nil? last-check)
+        (-> last-check
+            time/parse-iso8601
+            time/delta-minutes
+            (> 5)))))
+
+
+(defn credential-can-op-check?
+  [credential]
+  (general-utils/can-operation? "check" credential))
+
+
+(defn credential-need-check?
+  [credential]
+  (boolean
+    (and
+      credential
+      (credential-can-op-check? credential)
+      (or (not (credential-status-valid credential))
+          (credential-is-outdated? credential)))))